fix(web): validate browse paths on the server
