+++ packages/web/src/app/(app)/browse/hooks/useBrowseParams.ts
@@ -1,22 +1,13 @@
-import { notFound, usePathname } from "next/navigation";
-import { useMemo } from "react";
-import { getBrowseParamsFromPathParam } from "./utils";
+import { createContext, useContext } from "react";
+import type { BrowseProps } from "./utils";
+
+export const BrowseParamsContext = createContext<BrowseProps | null>(null);
 
 export const useBrowseParams = () => {
-    const pathname = usePathname();
+    const browseParams = useContext(BrowseParamsContext);
+    if (!browseParams) {
+        throw new Error('useBrowseParams must be used within a BrowseParamsContext provider');
+    }
 
-    return useMemo(() => {
-        const startIndex = pathname.indexOf('/browse/');
-        if (startIndex === -1) {
-            notFound();
-        }
-
-        const rawPath = pathname.substring(startIndex + '/browse/'.length);
-        const browseParams = getBrowseParamsFromPathParam(rawPath);
-        if (!browseParams) {
-            notFound();
-        }
-
-        return browseParams;
-    }, [pathname]);
+    return browseParams;
 }
